add back ui/action_collection and create api/action_collection
