+++ lib/flipper/api/v1/actions/boolean_gate.rb
@@ -8,16 +8,16 @@ module Flipper
           route %r{api/v1/features/[^/]*/(enable|disable)/?\Z}
           
           def put
-            feature_name = Rack::Utils.unescape(route_parts[-2])
+            feature_name = Rack::Utils.unescape(path_parts[-2])
             feature = flipper[feature_name.to_sym]
-            action = Rack::Utils.unescape(route_parts.last)
+            action = Rack::Utils.unescape(path_parts.last)
             feature.send(action)
-            json_response({feature: feature})
+            json_response({}, 204)
           end
 
           private
 
-          def route_parts
+          def path_parts
             request.path.split("/")
           end
         end
